Function snowbridge_outbound_queue_merkle_tree::merkle_root

pub fn merkle_root<H, I>(leaves: I) -> H256
where
    H: Hash<Output = H256>,
    I: Iterator<Item = H256>,

Construct a root hash of a Binary Merkle Tree created from given leaves.

See crate-level docs for details about Merkle Tree construction.

In case an empty list of leaves is passed the function returns a 0-filled hash.