Function snowbridge_outbound_queue_merkle_tree::verify_proof

pub fn verify_proof<'a, H, P, L>(
    root: &'a H256,
    proof: P,
    number_of_leaves: u64,
    leaf_index: u64,
    leaf: L,
) -> bool
where
    H: Hash<Output = H256>,
    P: IntoIterator<Item = H256>,
    L: Into<Leaf<'a>>,

Verify Merkle Proof correctness versus given root hash.

The proof is NOT expected to contain leaf hash as the first element, but only all adjacent nodes required to eventually by process of concatenating and hashing end up with given root hash.

The proof must not contain the root hash.